Added custom rolloff and data reading/seeking functionality to Sound class.
I'm not closing the relevant issue just yed, because i got a lot of fmod errors when i tryed to test the functions (Update error during a read_data call and a seeking error during a seek). The wrapper code is probably correct, but i would appreciate some more testing.